clean up code and create use hook
